@@ -2,95 +2,45 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.SqlClient;
+using System.Linq;
+using System.Threading.Tasks;
+using MongoDB.Bson;
+using MongoDB.Driver;
 
-namespace EscapeFromTheWoods
-{
-    public class DBwriter
-    {
-        private string connectionString;
+namespace EscapeFromTheWoods;
 
-        public DBwriter(string connectionString)
-        {
-            this.connectionString = connectionString;
-        }
+public class DBwriter {
+    private string _connectionString;
+    private IMongoClient mongoClient;
+    private IMongoDatabase mongoDB;
 
-        private SqlConnection getConnection()
-        {
-            SqlConnection connection = new SqlConnection(connectionString);
-            return connection;
-        }
-        public void WriteWoodRecords(List<DBWoodRecord> data)
-        {
-            SqlConnection connection = getConnection();
-            string query = "INSERT INTO dbo.WoodRecords (woodID,treeID,x,y) VALUES(@woodID,@treeID,@x,@y)";
-            using (SqlCommand command = connection.CreateCommand())
-            {
-                connection.Open();
-                try
-                {
-                    command.Parameters.Add(new SqlParameter("@woodID", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@treeID", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@x", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@y", SqlDbType.Int));
+    public DBwriter(string connectionString) {
+        _connectionString = connectionString;
+        mongoClient = new MongoClient(connectionString);
+        mongoDB = mongoClient.GetDatabase("EscapeFromTheWoods");
+    }
+    public async Task WriteWoodRecordsAsync(List<DBWoodRecord> data) {
+        var collection = mongoDB.GetCollection<BsonDocument>("Wood");
+        var document = data.Select(x => new BsonDocument
+        { { "woodID", x.woodID },
+          { "treeID", x.treeID },
+          { "x", x.x },
+          { "y", x.y }, });
+        await collection.InsertManyAsync(document);
+    }
 
-                    command.CommandText = query;
-                    foreach (var x in data)
-                    {
-                        command.Parameters["@woodID"].Value = x.woodID;
-                        command.Parameters["@treeID"].Value = x.treeID;
-                        command.Parameters["@x"].Value = x.x;
-                        command.Parameters["@y"].Value = x.y;
-                        command.ExecuteNonQuery();
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex);
-                }
-                finally
-                {
-                    connection.Close();
-                }
-            }
-        }
-        public void WriteMonkeyRecords(List<DBMonkeyRecord> data)
-        {
-            SqlConnection connection = getConnection();
-            string query = "INSERT INTO dbo.MonkeyRecords (monkeyID,monkeyName,woodID, seqNr,treeID,x,y) VALUES(@monkeyID,@monkeyName,@woodID,@seqNr,@treeID,@x,@y)";
-            using (SqlCommand command = connection.CreateCommand())
-            {
-                connection.Open();
-                try
-                {
-                    command.Parameters.Add(new SqlParameter("@monkeyID", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@monkeyName", SqlDbType.NVarChar));
-                    command.Parameters.Add(new SqlParameter("@woodID", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@seqNr", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@treeID", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@x", SqlDbType.Int));
-                    command.Parameters.Add(new SqlParameter("@y", SqlDbType.Int));
-
-                    command.CommandText = query;
-                    foreach (var x in data) {
-                        command.Parameters["@monkeyID"].Value = x.monkeyID;
-                        command.Parameters["@monkeyName"].Value = x.monkeyName;
-                        command.Parameters["@woodID"].Value = x.woodID;
-                        command.Parameters["@seqNr"].Value = x.seqNr;
-                        command.Parameters["@treeID"].Value = x.treeID;
-                        command.Parameters["@x"].Value = x.x;
-                        command.Parameters["@y"].Value = x.y;
-                        command.ExecuteNonQuery();
-                    }
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex);
-                }
-                finally
-                {
-                    connection.Close();
-                }
-            }
-        }
+    public async Task WriteMonkeyRecordsAsyncDb(List<DBMonkeyRecord> data) {
+        var collection = mongoDB.GetCollection<BsonDocument>("Monkey");
+        var document = data.Select(x => new BsonDocument
+        { 
+            { "monkeyID", x.monkeyID },
+            { "monkeyName", x.monkeyName },
+            { "woodID", x.woodID },
+            { "seqNr", x.seqNr },
+            { "treeID", x.treeID },
+            { "x", x.x },
+            { "y", x.y }
+        });
+        await collection.InsertManyAsync(document);
     }
 }
