@@ -1,29 +1,30 @@
 ﻿using System;
 using System.Diagnostics;
+using System.Threading.Tasks;
 
-namespace EscapeFromTheWoods
-{
-    class Program
-    {
-        static void Main(string[] args)
-        {
+namespace EscapeFromTheWoods {
+    class Program {
+        static async Task Main(string[] args) {
             Stopwatch stopwatch = new Stopwatch();
             stopwatch.Start();
             Console.WriteLine("Hello World!");
-            string connectionString = @"Data Source=lenovo-pc\sqlexpress;Initial Catalog=monkeys;Integrated Security=True";
+            // string connectionString =
+            //     @"Data Source=.\SQLEXPRESS;Initial Catalog=ETW;Integrated Security=True;TrustServerCertificate=True";
+            string connectionString = "mongodb://localhost:27017";
             DBwriter db = new DBwriter(connectionString);
 
-            string path = @"C:\NET\monkeys";
+            string path =
+                @"D:\CursusPB\Alle_opdrachte\EscapeFromTheWoods\EscapeFromTheWoodsToRefactor\EscapeFromTheWoodsToRefactor\NET\monkeys";
             Map m1 = new Map(0, 500, 0, 500);
-            Wood w1 = WoodBuilder.GetWood(500, m1, path,db);
+            Wood w1 = WoodBuilder.GetWood(500, m1, path, db);
             w1.PlaceMonkey("Alice", IDgenerator.GetMonkeyID());
             w1.PlaceMonkey("Janice", IDgenerator.GetMonkeyID());
             w1.PlaceMonkey("Toby", IDgenerator.GetMonkeyID());
             w1.PlaceMonkey("Mindy", IDgenerator.GetMonkeyID());
             w1.PlaceMonkey("Jos", IDgenerator.GetMonkeyID());
-            
+
             Map m2 = new Map(0, 200, 0, 400);
-            Wood w2 = WoodBuilder.GetWood(2500, m2, path,db);
+            Wood w2 = WoodBuilder.GetWood(2500, m2, path, db);
             w2.PlaceMonkey("Tom", IDgenerator.GetMonkeyID());
             w2.PlaceMonkey("Jerry", IDgenerator.GetMonkeyID());
             w2.PlaceMonkey("Tiffany", IDgenerator.GetMonkeyID());
@@ -31,19 +32,24 @@ namespace EscapeFromTheWoods
             w2.PlaceMonkey("Jebus", IDgenerator.GetMonkeyID());
 
             Map m3 = new Map(0, 400, 0, 400);
-            Wood w3 = WoodBuilder.GetWood(2000, m3, path,db);
+            Wood w3 = WoodBuilder.GetWood(2000, m3, path, db);
             w3.PlaceMonkey("Kelly", IDgenerator.GetMonkeyID());
             w3.PlaceMonkey("Kenji", IDgenerator.GetMonkeyID());
             w3.PlaceMonkey("Kobe", IDgenerator.GetMonkeyID());
             w3.PlaceMonkey("Kendra", IDgenerator.GetMonkeyID());
 
-            w1.WriteWoodToDB();
-            w2.WriteWoodToDB();
-            w3.WriteWoodToDB();
-            w1.Escape();
-            w2.Escape();
-            w3.Escape();
-            
+
+            await Task.WhenAll(new[]
+            { w1.WriteWoodToDbAsync(),
+              w2.WriteWoodToDbAsync(),
+              w3.WriteWoodToDbAsync() });
+
+            await Task.WhenAll(new[]
+            { w1.EscapeAsync(m1),
+              w2.EscapeAsync(m2),
+              w3.EscapeAsync(m3), });
+
+
             stopwatch.Stop();
             // Write result.
             Console.WriteLine("Time elapsed: {0}", stopwatch.Elapsed);
